refactor: isolate workspace composer state and keyboard logic
Extract the workspace composer into a dedicated component so draft keystrokes no longer update parent state on every input event. This keeps submit, textarea sizing, and composer error handling encapsulated while preserving the existing UX and keyboard semantics.

- add new WorkspaceComposer component with local draft state, autosize/corner-radius layout metrics, utility/thinking menus, and cmd+enter submission behavior
- expose a minimal imperative handle for quote insertion, global type-to-focus behavior, and restoring draft text after send failures
- replace parent-level draft/setDraft wiring with stable callback props and primitive status flags for streaming/disabled/error states
- keep parent composer padding/collapse coordination via onComposerPaddingChange without reintroducing per-keystroke parent renders
- update WorkspaceClient tests for restored-error behavior and add explicit cmd+enter vs shift+enter keyboard submission coverage
